+++ src/ArduinoJson/Array/ElementProxy.hpp
@@ -13,14 +13,18 @@
 #endif
 
 namespace ARDUINOJSON_NAMESPACE {
-class ArraySubscript : public VariantOperators<ArraySubscript>,
-                       public Visitable {
+
+template <typename TArray>
+class ElementProxy : public VariantOperators<ElementProxy<TArray> >,
+                     public Visitable {
+  typedef ElementProxy<TArray> this_type;
+
  public:
-  FORCE_INLINE ArraySubscript(ArrayRef array, size_t index)
+  FORCE_INLINE ElementProxy(TArray array, size_t index)
       : _array(array), _index(index) {}
 
-  FORCE_INLINE ArraySubscript& operator=(const ArraySubscript& src) {
-    get_impl().set(src.as<VariantConstRef>());
+  FORCE_INLINE this_type& operator=(const this_type& src) {
+    getElement().set(src.as<VariantConstRef>());
     return *this;
   }
 
@@ -30,36 +34,36 @@ class ArraySubscript : public VariantOperators<ArraySubscript>,
   // TValue = bool, long, int, short, float, double, serialized, VariantRef,
   //          std::string, String, ArrayRef, ObjectRef
   template <typename T>
-  FORCE_INLINE ArraySubscript& operator=(const T& src) {
-    get_impl().set(src);
+  FORCE_INLINE this_type& operator=(const T& src) {
+    getElement().set(src);
     return *this;
   }
   //
   // operator=(TValue)
   // TValue = char*, const char*, const __FlashStringHelper*
   template <typename T>
-  FORCE_INLINE ArraySubscript& operator=(T* src) {
-    get_impl().set(src);
+  FORCE_INLINE this_type& operator=(T* src) {
+    getElement().set(src);
     return *this;
   }
 
   FORCE_INLINE bool isNull() const {
-    return get_impl().isNull();
+    return getElement().isNull();
   }
 
   template <typename T>
   FORCE_INLINE typename VariantAs<T>::type as() const {
-    return get_impl().as<T>();
+    return getElement().template as<T>();
   }
 
   template <typename T>
   FORCE_INLINE bool is() const {
-    return get_impl().is<T>();
+    return getElement().template is<T>();
   }
 
   template <typename T>
   FORCE_INLINE typename VariantTo<T>::type to() const {
-    return get_impl().to<T>();
+    return getElement().template to<T>();
   }
 
   // Replaces the value
@@ -69,42 +73,65 @@ class ArraySubscript : public VariantOperators<ArraySubscript>,
   //          std::string, String, ArrayRef, ObjectRef
   template <typename TValue>
   FORCE_INLINE bool set(const TValue& value) const {
-    return get_impl().set(value);
+    return getElement().set(value);
   }
   //
   // bool set(TValue)
   // TValue = char*, const char*, const __FlashStringHelper*
   template <typename TValue>
   FORCE_INLINE bool set(TValue* value) const {
-    return get_impl().set(value);
+    return getElement().set(value);
   }
 
   template <typename Visitor>
   void accept(Visitor& visitor) const {
-    return get_impl().accept(visitor);
+    return getElement().accept(visitor);
   }
 
   FORCE_INLINE size_t size() const {
-    return get_impl().size();
+    return getElement().size();
+  }
+
+  template <typename TNestedKey>
+  VariantRef get(TNestedKey* key) const {
+    return getElement().get(key);
+  }
+
+  template <typename TNestedKey>
+  VariantRef get(const TNestedKey& key) const {
+    return getElement().get(key);
+  }
+
+  template <typename TNestedKey>
+  VariantRef getOrCreate(TNestedKey* key) const {
+    return getElement().getOrCreate(key);
+  }
+
+  template <typename TNestedKey>
+  VariantRef getOrCreate(const TNestedKey& key) const {
+    return getElement().getOrCreate(key);
+  }
+
+  using ArrayShortcuts<ElementProxy>::add;
+  VariantRef add() const {
+    return getElement().add();
   }
 
  private:
-  FORCE_INLINE VariantRef get_impl() const {
+  FORCE_INLINE VariantRef getElement() const {
     return _array.get(_index);
   }
 
-  ArrayRef _array;
+  TArray _array;
   const size_t _index;
 };
 
-template <typename TImpl>
-inline ArraySubscript VariantSubscripts<TImpl>::operator[](size_t index) const {
-  return impl()->template as<ArrayRef>()[index];
+template <typename TArray>
+inline ElementProxy<const TArray&> ArrayShortcuts<TArray>::operator[](
+    size_t index) const {
+  return ElementProxy<const TArray&>(*impl(), index);
 }
 
-inline ArraySubscript ArrayRef::operator[](size_t index) const {
-  return ArraySubscript(*this, index);
-}
 }  // namespace ARDUINOJSON_NAMESPACE
 
 #ifdef _MSC_VER